[VM][FM7] gcc-5 -O3 is too radical to build.

+++ source/src/vm/fm7/fm7_display.h
@@ -58,7 +58,7 @@ class DISPLAY: public DEVICE
 	void reset_cpuonly(void);
    
 #if defined(_FM77AV_VARIANTS)
-	void alu_write_cmdreg(uint8 val);
+	void alu_write_cmdreg(uint32 val);
 	void alu_write_logical_color(uint8 val);
 	void alu_write_mask_reg(uint8 val);
 	void alu_write_cmpdata_reg(int addr, uint8 val);
